Renamed ActionManager to ActionLoop
