+++ vizro-core/src/vizro/actions/_action_loop/_get_action_loop_components.py
